requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.4
    # via flask-migrate
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
apscheduler==3.11.0
    # via -r requirements.in
authlib==1.6.1
    # via -r requirements.in
beautifulsoup4==4.13.4
    # via -r requirements.in
black==25.1.0
    # via -r requirements.in
blinker==1.9.0
    # via
    #   flask
    #   flask-mail
boto3==1.39.14
    # via -r requirements.in
botocore==1.39.14
    # via
    #   boto3
    #   s3transfer
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   flask
coverage[toml]==7.10.0
    # via
    #   -r requirements.in
    #   pytest-cov
cryptography==45.0.5
    # via authlib
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via -r requirements.in
execnet==2.1.1
    # via pytest-xdist
feedparser==6.0.11
    # via -r requirements.in
filelock==3.18.0
    # via virtualenv
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-mail
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
    #   pytest-flask
flask-login==0.6.3
    # via -r requirements.in
flask-mail==0.10.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.2.2
    # via -r requirements.in
greenlet==3.2.3
    # via sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
jiter==0.10.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mypy==1.17.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
newsapi-python==0.2.7
    # via -r requirements.in
nodeenv==1.9.1
    # via pre-commit
openai==1.97.1
    # via -r requirements.in
packaging==25.0
    # via
    #   black
    #   gunicorn
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==11.3.0
    # via -r requirements.in
platformdirs==4.3.8
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.2.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-flask
    #   pytest-xdist
pytest-cov==6.2.1
    # via -r requirements.in
pytest-flask==1.3.0
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via -r requirements.in
pyyaml==6.0.2
    # via pre-commit
requests==2.32.4
    # via
    #   -r requirements.in
    #   newsapi-python
ruff==0.12.7
    # via -r requirements.in
s3transfer==0.13.1
    # via boto3
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.41
    # via
    #   alembic
    #   flask-sqlalchemy
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   alembic
    #   anyio
    #   beautifulsoup4
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via
    #   botocore
    #   requests
virtualenv==20.32.0
    # via pre-commit
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
    #   pytest-flask
wtforms==3.2.1
    # via flask-wtf
